wip: rename types file
